requirements.txt
# --- Cœur de l'application ---
streamlit==1.28.0

# --- Manipulation de données ---
pandas==2.0.3
openpyxl==3.1.2
numpy==1.24.3

# --- Intelligence Artificielle ---
scikit-learn==1.3.0
joblib==1.3.2
opencv-python-headless==4.8.1.78

# --- Visualisation & Cartographie (Nouveau) ---
matplotlib==3.7.2
seaborn==0.12.2
folium>=0.14.0
streamlit-folium>=0.15.0
Pillow>=10.0.0